Some fixes that apply to legacy compilers or GLIBC.

+++ src/libxsmm.c
@@ -196,7 +196,7 @@ LIBXSMM_RETARGETABLE LIBXSMM_VISIBILITY_INTERNAL LIBXSMM_LOCK_TYPE internal_regl
 #   define INTERNAL_FIND_CODE_READ(CODE, DST) DST = __sync_or_and_fetch(&(CODE)->function.pmm, 0)
 #   define INTERNAL_FIND_CODE_WRITE(CODE, SRC) { \
       /*const*/void* old = (CODE)->function.pmm; \
-      while (!__sync_bool_compare_and_swap(&(CODE)->function.pmm, old, SRC)) { \
+      while (!__sync_bool_compare_and_swap(&(CODE)->function.pmm, (uintptr_t)old, (uintptr_t)(SRC))) { \
         old = (CODE)->function.pmm; \
       } \
     }
@@ -595,7 +595,7 @@ LIBXSMM_INLINE LIBXSMM_RETARGETABLE internal_regentry* internal_init(void)
 # else
           {
             internal_regentry* old = internal_registry;
-            while (!__sync_bool_compare_and_swap(&internal_registry, old, result)) old = internal_registry;
+            while (!__sync_bool_compare_and_swap(&internal_registry, (uintptr_t)old, (uintptr_t)result)) old = internal_registry;
           }
 # endif
 #elif (defined(_REENTRANT) || defined(LIBXSMM_OPENMP)) && defined(_WIN32)
